Moved select many components to separate file.
