+++ app/src/main/java/com/koleff/kare_android/ui/compose/screen/SearchExercisesScreen.kt
@@ -1,44 +1,28 @@
 package com.koleff.kare_android.ui.compose.screen
 
 import android.util.Log
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.koleff.kare_android.common.MockupDataGeneratorV2
 import com.koleff.kare_android.common.navigation.Destination
 import com.koleff.kare_android.common.navigation.NavigationEvent
 import com.koleff.kare_android.data.model.dto.ExerciseDto
@@ -46,12 +30,13 @@ import com.koleff.kare_android.data.model.response.base_response.KareError
 import com.koleff.kare_android.ui.compose.components.LoadingWheel
 import com.koleff.kare_android.ui.compose.components.SearchBar
 import com.koleff.kare_android.ui.compose.components.SearchExercisesList
+import com.koleff.kare_android.ui.compose.components.SelectedExercisesRow
+import com.koleff.kare_android.ui.compose.components.SubmitExercisesRow
 import com.koleff.kare_android.ui.compose.components.navigation_components.scaffolds.SearchListScaffold
 import com.koleff.kare_android.ui.compose.dialogs.DuplicateExercisesFoundDialog
 import com.koleff.kare_android.ui.compose.dialogs.ErrorDialog
 import com.koleff.kare_android.ui.event.OnMultipleExercisesUpdateEvent
 import com.koleff.kare_android.ui.state.BaseState
-import com.koleff.kare_android.ui.theme.LocalExtendedColorScheme
 import com.koleff.kare_android.ui.view_model.SearchExercisesViewModel
 
 @Composable
@@ -221,132 +206,6 @@ fun SearchExercisesScreen(
                 }
             }
         }
-    }
-}
-
-@Composable
-fun SelectedExercisesRow(selectedExercises: List<ExerciseDto>) {
-    val backgroundColor = MaterialTheme.colorScheme.tertiary
-    val outlineColor = MaterialTheme.colorScheme.outlineVariant
-
-    val textColor = MaterialTheme.colorScheme.onSurface
-    val textStyle = MaterialTheme.typography.titleMedium.copy(
-        color = textColor
-    )
-
-    val cornerSize = 16.dp
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(60.dp)
-            .padding(horizontal = 16.dp)
-            .clip(RoundedCornerShape(cornerSize))
-            .background(
-                color = backgroundColor,
-                shape = RoundedCornerShape(cornerSize)
-            )
-            .border(
-                border = BorderStroke(2.dp, color = outlineColor),
-                shape = RoundedCornerShape(cornerSize)
-            ),
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = "${selectedExercises.size} exercises selected.",
-            style = textStyle,
-            maxLines = 2,
-            overflow = TextOverflow.Ellipsis
-        )
-    }
-}
-
-@Composable
-fun UpdateExercisesRow(
-    modifier: Modifier = Modifier,
-    text: String,
-    backgroundColor: Color,
-    onAction: () -> Unit
-) {
-    val outlineColor = MaterialTheme.colorScheme.outlineVariant
-
-    val textColor = MaterialTheme.colorScheme.onSurface
-    val textStyle = MaterialTheme.typography.titleMedium.copy(
-        color = textColor
-    )
-
-    val cornerSize = 16.dp
-
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(60.dp)
-            .clip(RoundedCornerShape(cornerSize))
-            .background(
-                color = backgroundColor,
-                shape = RoundedCornerShape(cornerSize)
-            )
-            .border(
-                border = BorderStroke(2.dp, color = outlineColor),
-                shape = RoundedCornerShape(cornerSize)
-            )
-            .clickable { onAction() },
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = text,
-            style = textStyle,
-            overflow = TextOverflow.Ellipsis
-        )
-    }
-}
-
-@Composable
-fun SubmitExercisesRow(modifier: Modifier = Modifier, onSubmit: () -> Unit) {
-    UpdateExercisesRow(
-        modifier = modifier,
-        text = "Add exercises",
-        backgroundColor = LocalExtendedColorScheme.current.workoutBannerColors.selectButtonColor,
-        onAction = onSubmit
-    )
-}
-
-@Composable
-fun DeleteExercisesRow(modifier: Modifier = Modifier, onDelete: () -> Unit) {
-    UpdateExercisesRow(
-        modifier = modifier,
-        text = "Delete exercises",
-        backgroundColor = LocalExtendedColorScheme.current.workoutBannerColors.deleteButtonColor,
-        onAction = onDelete
-    )
-}
-
-@Preview
-@PreviewLightDark
-@Composable
-private fun SelectedExercisesRowPreview() {
-    val selectedExercises = listOf<ExerciseDto>()
-
-    SelectedExercisesRow(selectedExercises)
-}
-
-@Preview
-@PreviewLightDark
-@Composable
-private fun DeleteExercisesRowPreview() {
-    DeleteExercisesRow() {
-
-    }
-}
-
-@Preview
-@PreviewLightDark
-@Composable
-private fun SubmitExercisesRowPreview() {
-    SubmitExercisesRow() {
-
     }
 }
